rename: 클래스 위치 변경
패키지 오류인지 확인하기 위해서
