+++ src/main/java/bridge/Application.java
@@ -1,10 +1,5 @@
 package bridge;
 
-import bridge.service.BridgeGame;
-import bridge.service.PlayerSetting;
-import bridge.view.InputView;
-import bridge.view.OutputView;
-
 public class Application {
     private static final BridgeGame bridgeGame = new BridgeGame();
     private static final InputView inputView = new InputView();
