+++ src/main/java/bridge/BridgeGame.java
@@ -1,9 +1,4 @@
-package bridge.service;
-
-import bridge.domain.GameData;
-import bridge.dto.HistoryMapDto;
-import bridge.dto.PlayerTryCountDto;
-import bridge.exception.InputException;
+package bridge;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
