+++ src/main/java/bridge/GameData.java
@@ -1,6 +1,4 @@
-package bridge.domain;
-
-import bridge.service.PlayerCommend;
+package bridge;
 
 import java.util.ArrayList;
 import java.util.List;
