+++ src/main/java/bridge/InputException.java
@@ -1,6 +1,6 @@
-package bridge.exception;
+package bridge;
 
-import static bridge.service.PlayerCommend.*;
+import static bridge.PlayerCommend.*;
 
 public class InputException extends IllegalArgumentException {
     private static final int MAX_SIZE = 20;
